RAPRITHPRA/sunita willams/sunita williams/AI-CHATBOT/requirements.txt
# No external dependencies required
# This package uses only the Python standard library 

# Core Dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
colorama==0.4.6

# Google AI and APIs
google-generativeai>=0.4.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Machine Learning and AI
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
tensorflow==2.13.0
joblib==1.3.2

# Text Processing and NLP
googletrans==4.0.0rc1
PyPDF2==3.0.1

# Audio and Speech
gtts==2.4.0
SpeechRecognition==3.10.0
pygame==2.5.2
pyaudio==0.2.11

# Image Processing
Pillow==10.0.1
opencv-python==4.8.1.78

# YouTube and Video
youtube-transcript-api==0.6.1

# Web and HTTP
requests==2.31.0
urllib3==2.0.7

# Data Processing
pandas==2.0.3

# Development and Testing
pytest==7.4.2
pytest-flask==1.3.0

# Database
sqlite3

# Optional: For better voice recognition (Windows)
# pywin32==306  # Uncomment on Windows

# Optional: For better audio processing
# pydub==0.25.1  # Uncomment if needed 